Mvh/integrate to scribe

+++ src/PdfGen.js
@@ -5,12 +5,20 @@ const pages = require('../resources/pages.json');
 const sizes = require('../resources/sizes.json');
 const handlers = require('./sectionHandlerClasses');
 const {sectionHandlerLookup} = require('./sectionHandlerLookup');
+const {resolvePath} = require('./helpers');
 const fse = require("fs-extra");
+const path = require("path");
+const os = require("os");
 
 class PdfGen {
 
-    constructor(options) {
+    constructor(options, doPdfCallback = null) {
+        const errors = PdfGen.validateConfig(options, true);
+        if (errors.length > 0) {
+            throw new Error(`Validation errors for config file:\n${errors.map(e => `  - ${e}`).join('\n')}`);
+        }
         this.options = options;
+        this.doPdfCallback = doPdfCallback;
     }
 
     static handlerInfo() {
@@ -32,22 +40,24 @@ class PdfGen {
     }
 
     async doPdf() {
-        const options = this.options;
-        // Check that output file will not accidentally overwrite an existing file
-        if (fse.pathExistsSync(options.output) && !options.forceOverwrite) {
-            throw new Error(`Output path '${options.output}' would be overwritten but 'forceOverwrite' flag is not set.`);
+        if (!this.options.global.outputPath) {
+            throw new Error("Must provide an outputPath in doPdf()");
         }
-
-        // Attempt to read config file and add JSON to options
-        if (!(fse.pathExistsSync(options.config)) || !(fse.lstatSync(options.config).isFile())) {
-            throw new Error(`Config file path '${options.config}' does not exist or is not a file`);
-        }
-        try {
-            options.configContent = fse.readJsonSync(options.config);
-        } catch (err) {
-            throw new Error(`Could not read config file ${options.config} as JSON: ${err.message}`);
-        }
-
+        const options = {
+            verbose: this.options.global.verbose || false,
+            workingDir: this.options.global.workingDir,
+            resourcesDir: this.options.global.resourcesDir,
+            pdfPath: path.join(this.options.global.workingDir, "pdf"),
+            htmlPath: path.join(this.options.global.workingDir, "html", "pages"),
+            manifestPath: path.join(this.options.global.workingDir, "manifest.json"),
+            steps: ["originate", "assemble"],
+            pageFormat: pages[this.options.global.pages],
+            fonts: fonts[this.options.global.fonts],
+            fontSizes: sizes[this.options.global.sizes],
+            configContent: this.options,
+            output: this.options.global.outputPath
+        };
+        this.options = options;
         // In CLEAR mode, delete working dir and exit
         if (options.steps.includes("clear")) {
             options.verbose && console.log("** CLEAR **");
@@ -62,7 +72,16 @@ class PdfGen {
 
         if (options.steps.includes("originate")) {
             options.verbose && console.log("** ORIGINATE **");
+            this.doPdfCallback && this.doPdfCallback({
+                type: "step",
+                level: 0,
+                msg: "Originating Content",
+                args: ["originate"]
+            });
             if (fse.pathExistsSync(options.workingDir)) {
+                if (!/[A-Za-z]{2}/.test(options.workingDir)) {
+                    throw new Error(`Working dir '${options.workingDir} looks dangerous to delete - quitting'`);
+                }
                 options.verbose && console.log(`   Deleting working dir ${options.workingDir}`);
                 fse.removeSync(options.workingDir);
             } else {
@@ -73,6 +92,12 @@ class PdfGen {
         }
         if (options.steps.includes("assemble")) {
             options.verbose && console.log("** ASSEMBLE **");
+            this.doPdfCallback && this.doPdfCallback({
+                type: "step",
+                level: 0,
+                msg: "Assembling Content",
+                args: ["assemble"]
+            });
             if (!fse.pathExistsSync(options.manifestPath)) {
                 throw new Error("Cannot run assemble without first originating content");
             }
@@ -80,7 +105,7 @@ class PdfGen {
         }
     }
 
-    validateConfig(configOb) {
+    static validateConfig(configOb, checkPaths = false) {
         const ret = [];
         let skip = false;
         // Top-level checks, abort on first error
@@ -107,10 +132,13 @@ class PdfGen {
         // globals checks, abort at end on error
         if (!skip) {
             const globalSpecKeys = Object.fromEntries(
-                Object.entries(this.constructor.pageInfo())
+                Object.entries(PdfGen.pageInfo())
                     .map(([k, v]) => [k, Object.keys(v)])
             );
             for (const globalKey of Object.keys(configOb.global)) {
+                if (["verbose", "workingDir", "outputPath"].includes(globalKey)) {
+                    continue;
+                }
                 if (!globalSpecKeys[globalKey]) {
                     ret.push(`Unexpected global key '${globalKey}'`);
                     skip = true;
@@ -136,7 +164,9 @@ class PdfGen {
                 while (sectionN < sections.length && !skip) {
                     const section = sections[sectionN];
                     if (sectionHandlerLookup[section.type]) { // A section
-                        if (!this.validateSection(section, null, ret, sectionN)) {
+                        const sectionErrors = this.validateSection(section, null, ret, sectionN, checkPaths);
+                        if (sectionErrors.length > 0) {
+                            sectionErrors.forEach(e => ret.push(e));
                             skip = true;
                             break;
                         }
@@ -148,10 +178,15 @@ class PdfGen {
                             }
                             wrapperOnly[key] = value;
                         }
+                        if (!wrapperOnly["ranges"] || !Array.isArray(wrapperOnly["ranges"]) || wrapperOnly["ranges"].length === 0) {
+                            ret.push(`Wrapper type '${section.type} requires ranges at '${section.id ? ` near section ${section.id} ` : ""}(#${sectionN})`);
+                            skip = true;
+                            break;
+                        }
                         for (const subSection of section.sections) {
-                            if (!this.validateSection(subSection, wrapperOnly, ret, sectionN)) {
-                             skip = true;
-                             break;
+                            if (!PdfGen.validateSection(subSection, wrapperOnly, ret, sectionN, checkPaths)) {
+                                skip = true;
+                                break;
                             }
                         }
                     } else {
@@ -166,76 +201,120 @@ class PdfGen {
         return ret;
     }
 
-    validateSection(section, wrapper, errors, sectionN) {
+    static validateSection(section, wrapper, errors, sectionN, checkPaths = false) {
         if (!section.id) {
             errors.push(`Section has no id (#${sectionN})`);
-            return false;
+            return errors;
         }
         if (!section.type) {
             errors.push(`Section '${section.id}' has no type (#${sectionN})`);
-            return false;
+            return errors;
         }
         const sectionHandler = sectionHandlerLookup[section.type];
         if (!sectionHandler) {
             errors.push(`Unknown section type '${section.type}' near section '${section.id}' (#${sectionN})`);
-            return false;
+            return errors;
         }
         const signature = sectionHandler.signature();
         // console.log(JSON.stringify(signature, null, 2));
         if (signature.requiresWrapper.length > 0) {
             if (!wrapper) {
                 errors.push(`No wrapper for section '${section.id}' of type '${section.type}' (expected '${signature.requiresWrapper[0]}) (#${sectionN})'`);
-                return false;
+                return errors;
             }
             if (wrapper.type !== `${signature.requiresWrapper[0]}Wrapper`) {
                 errors.push(`Expected wrapper of type '${signature.requiresWrapper[0]}' for section '${section.id}' but found '${wrapper.type}' (#${sectionN})`);
-                return false;
+                return errors;
             }
         }
-        const signatureFieldIds = signature.fields.map(f => f.id);
-        for (const sectionFieldId of Object.keys(section.content)) {
+        if (!section.content) {
+            errors.push(`No content for section '${section.id}' (#${sectionN})`);
+            return errors;
+        }
+        return PdfGen.validateSectionFields(
+            {
+                signatureFields: signature.fields,
+                sectionId: section.id,
+                sectionN,
+                sectionContent: section.content,
+                errors,
+                checkPaths
+            }
+        );
+    }
+
+    static validateSectionFields({signatureFields, sectionId, sectionN, sectionContent, errors, checkPaths = false}) {
+        const signatureFieldIds = signatureFields.map(f => f.id);
+        for (const sectionFieldId of Object.keys(sectionContent)) {
             if (!signatureFieldIds.includes(sectionFieldId)) {
-                errors.push(`Unexpected field '${sectionFieldId}' in Section '${section.id}' of type '${section.type}' (#${sectionN})`);
+                errors.push(`Unexpected field '${sectionFieldId}' in Section '${sectionId}' (#${sectionN})`);
             }
         }
-        for (const requiredFieldId of signature.fields.filter(f => f.nValues[0] > 0).map(f => f.id)) {
-            if (!section.content[requiredFieldId]) {
-                errors.push(`Missing field '${requiredFieldId}' in Section '${section.id}' of type '${section.type}' (#${sectionN})`);
+        for (const requiredFieldId of signatureFields.filter(f => f.nValues[0] > 0).map(f => f.id)) {
+            if (sectionContent[requiredFieldId] === undefined) {
+                errors.push(`Missing field '${requiredFieldId}' in Section '${sectionId}' (#${sectionN})`);
             }
         }
         if (errors.length > 0) {
-            return false;
+            return errors;
         }
         let foundError = false;
-        for (const [key, value] of Object.entries(section.content)) {
+        for (const [key, value] of Object.entries(sectionContent)) {
             if (
-                !this.validateSectionField(
+                !PdfGen.validateSectionField(
                     key,
                     value,
-                    signature.fields.filter(f => f.id === key)[0],
+                    signatureFields.filter(f => f.id === key)[0],
                     errors,
-                    section.id,
-                    sectionN
+                    sectionId,
+                    sectionN,
+                    checkPaths
                 )
             ) {
                 foundError = true;
             }
         }
-        return !foundError;
+        return errors;
+
     }
 
-    validateSectionField(fieldId, fieldContent, fieldSpec, errors, sectionId, sectionN) {
+    static validateSectionField(fieldId, fieldContent, fieldSpec, errors, sectionId, sectionN, checkPaths = false) {
         const normalizedContent = Array.isArray(fieldContent) ? fieldContent : [fieldContent];
         if (normalizedContent.length < fieldSpec.nValues[0] || normalizedContent.length > fieldSpec.nValues[1]) {
             errors.push(`${normalizedContent.length} values for field '${fieldId}' in Section '${sectionId}' (#${sectionN}) - expected ${fieldSpec.nValues[0]}-${fieldSpec.nValues[1]} value(s)`);
             return false;
         }
         if (fieldSpec.typeName) {
+            if (!["boolean", "number", "string", "obs", "tNotes", "translationText", "md", "juxta"].includes(fieldSpec.typeName)) {
+                errors.push(`Unknown typeName '${fieldSpec.typeName}' in Section '${sectionId}' (#${sectionN})`);
+                return false;
+            }
             if (fieldSpec.typeName === "boolean") {
                 const badValues = normalizedContent.filter(c => !!c !== c);
                 if (badValues.length > 0) {
                     errors.push(`${badValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) are not booleans`);
                     return false;
+                }
+            }
+            if (fieldSpec.typeName === "number") {
+                let badValues = normalizedContent.filter(c => typeof c !== "number");
+                if (badValues.length > 0) {
+                    errors.push(`${badValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) are not numbers`);
+                    return false;
+                }
+                if (fieldSpec.minValue) {
+                    badValues = normalizedContent.filter(c => c < fieldSpec.minValue);
+                    if (badValues.length > 0) {
+                        errors.push(`${badValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) are < minValue ${fieldSpec.minValue}`);
+                        return false;
+                    }
+                }
+                if (fieldSpec.maxValue) {
+                    badValues = normalizedContent.filter(c => c > fieldSpec.maxValue);
+                    if (badValues.length > 0) {
+                        errors.push(`${badValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) are > maxValue ${fieldSpec.maxValue}`);
+                        return false;
+                    }
                 }
             } else if (fieldSpec.typeName === "string") {
                 const badValues = normalizedContent.filter(c => typeof c !== "string");
@@ -243,11 +322,19 @@ class PdfGen {
                     errors.push(`${badValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) are not strings`);
                     return false;
                 }
-            } else if (["ob", "tNotes", "translationText"].includes(fieldSpec.typeName) && typeof fieldContent !== "string") {
+            } else if (["obs", "tNotes", "translationText", "md", "juxta"].includes(fieldSpec.typeName)) {
                 const badValues = normalizedContent.filter(c => typeof c !== "string");
                 if (badValues.length > 0) {
                     errors.push(`${badValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) are not path strings`);
                     return false;
+                }
+                if (checkPaths) {
+                    const unresolvedValues = normalizedContent
+                        .map(c => resolvePath(c))
+                        .filter(c => !fse.pathExistsSync(resolvePath(c)));
+                    if (unresolvedValues.length > 0) {
+                        errors.push(`${unresolvedValues.length} value(s) of field '${fieldId}' in Section '${sectionId}' (#${sectionN}) could not be resolved in the FS: ${unresolvedValues.map(p => `'${p}'`).join(',')}`);
+                    }
                 }
             }
         }
@@ -259,15 +346,36 @@ class PdfGen {
                 return false;
             }
         }
+        if (fieldSpec.typeSpec) {
+            let foundError = false;
+            for (const fieldContentItem of fieldContent) {
+                const fieldReport = PdfGen.validateSectionFields(
+                    {
+                        signatureFields: fieldSpec.typeSpec,
+                        sectionId,
+                        sectionN,
+                        sectionContent: fieldContentItem,
+                        errors,
+                        checkPaths: true
+                    }
+                );
+                if (fieldReport) {
+                    foundError = true;
+                }
+            }
+            if (foundError) {
+                return false;
+            }
+        }
         return true;
     }
 
     async originatePdfs() {
-        return await _originatePdfs(this.options);
+        return await _originatePdfs(this.options, this.doPdfCallback);
     }
 
     async assemblePdfs() {
-        return await _assemblePdfs(this.options);
+        return await _assemblePdfs(this.options, this.doPdfCallback);
     }
 }
